Feat: Add student grades import

+++ backend/src/main/java/fr/eseo/tauri/model/GradeType.java
@@ -19,7 +19,7 @@ public class GradeType {
     private String name;
 
     @JsonProperty
-    private Integer factor;
+    private Double factor;
 
     @JsonProperty
     private Boolean forGroup;
